parser.cを複数ファイルに分割
